Improve wording
Don't use word "behaviour" as it is clear that it is a behaviour.

Use verb "Filter" as prefix (not noun "filtering" as postfix) to comply
with other behaviours (e.g. "Filter by name..." from scm-api-plugin).

Use word "pattern" (not "tag") as users may be confused because git
itself uses tags.

Quote the patterns to make clear that "[" and "]" have to be contained
literally and are not just used as delimiters in the help itself.